Added 3 functions: determine_num_missiles; add_missiles; missiles_on_screen_check
These 3 functions are used for the logic  behind a round stystem. Each round allows a certain max number of missiles on screen at time. Also, when a certain number of missiles go off-screen, then new ones can spawn (while on the same round). If the player survives long enough, then the round should be incremented. (not sure how to implement this yet)